+++ src/assets/InputAsset.php
@@ -181,6 +181,7 @@ class InputAsset extends FieldAsset
                 'name' => Craft::t('site', $blockType->name),
                 'handle' => $blockType->handle,
                 'description' => $blockType->description,
+                'minBlocks' => $blockType->minBlocks,
                 'maxBlocks' => $blockType->maxBlocks,
                 'maxSiblingBlocks' => $blockType->maxSiblingBlocks,
                 'minChildBlocks' => $blockType->minChildBlocks,