+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 ## Unreleased (3.3.0)
 
 ### Added
+- Added the Min Blocks block type setting (added `benf\neo\models\BlockType::$minBlocks` and added the `minBlocks` column to the `neoblocktypes` table)
 - Added the Min Child Blocks block type setting (added `benf\neo\models\BlockType::$minChildBlocks` and added the `minChildBlocks` column to the `neoblocktypes` table)
+- Added `benf\neo\console\controllers\BlockTypesController::$setMinBlocks` (added the `--set-min-blocks` option to the `php craft neo/block-types/edit` console command)
 - Added `benf\neo\console\controllers\BlockTypesController::$setMinChildBlocks` (added the `--set-min-child-blocks` option to the `php craft neo/block-types/edit` console command)
+
+### Fixed
+- Fixed a bug when saving an element, where Neo fields weren't being validated on block types' Max Blocks settings, unless at least one of the Neo field's Max Top-Level Blocks and Max Levels settings were set
 
 ## 3.2.5 - 2022-08-03